whisper-service/requirements.txt
mlx-whisper>=0.4.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pyyaml>=6.0
soundfile>=0.12.0
numpy>=1.24.0
